chore: move some function to ha folder
